requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
twilio>=8.0.0
